creates telephone component to refactor the quotes form
