+++ src/index.css
@@ -135,6 +135,10 @@ height:100vh;
   height:100vh;
   }
 
+#telephoneQuestion{
+  height:100vh;
+  }
+
 /* ------------------------Mobile------------------------ */
 
 #bannerImgMob {
